requirements.txt
faster-whisper>=1.1.0
ctranslate2>=4.4.0
onnxruntime-gpu>=1.18.0
soundfile>=0.12.1
sounddevice>=0.4.6
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.1
